Models.jl Documentation

This simple package contains functions and lineshapes commonly used in vibrational polariton spectroscopic experiments. The main purpose is code reusability for my own projects, but on the off chance that it is useful to others, it is available. These functions are best used together with Optim.jl or the global optimization package Optimization.jl for fitting. It would be nice if models and fitting procedures were bundled together in the future, like the lmfit package available for Python (leveraging SciPy).

General Functions

Common functions and lineshapes for spectroscopy.

Models.exponential — Method

Exponential decay function.

$f(x; \tau, y_0) = f_0 e^{-x / \tau} + y_0$

https://en.wikipedia.org/wiki/Exponential_decay

Models.double_exponential — Method

Exponential decay function with two stages represented by two different time constants.

$f(x; f_1, \tau_1, f_2, \tau_2, y_0) = f_1 e^{-x / \tau_1} + f_2 e^{-x / \tau_2} + y_0$

Models.gaussian — Method

Gaussian function. $f(x; A, \mu, \sigma, y_0) = \frac{A}{\sigma \sqrt{2\pi}} e^{-(x - \mu)^2 / (2 \sigma^2)} + y_0$

https://en.wikipedia.org/wiki/Gaussian_function

Models.gaussian2d — Method

Two-dimensional Gaussian lineshape.

$f(x, y; A, x_0, \sigma_x, y_0, \sigma_y, z_0) = A \exp\left(-\left( \frac{(x - x_0)^2}{2 \sigma_x^2} + \frac{(y - y_0)^2}{2 \sigma_y^2} \right)\right) + z_0$

Models.lorentzian — Method

Lorentzian function with amplitude $A$, center frequency $\nu_0$, and full width at half maximum (FWHM) $2\gamma$.

$f(\nu; A, \nu_0, \gamma, y_0) = \frac{A}{\pi} \frac{\gamma}{(\nu - \nu_0)^2 + \gamma^2} + y_0$

https://en.wikipedia.org/wiki/Cauchy_distribution

Missing docstring.

Missing docstring for squared_errors(p). Check Documenter's build log for details.

Special Functions

These function are specific to cavity polariton research, where things like the cavity transmittance, material dielectric function, and cavity free spectral range (FSR) need to be calculated and modeled.

Models.cavity_transmittance — Method

Cavity transmittance as a function of frequency.

$T(\nu) = \frac{T^2 e^{-\alpha L}}{1 + R^2 e^{-2 \alpha L} - 2 R e^{-\alpha L} \cos(4\pi n L \nu + 2\phi)},$

where $\nu$ is the frequency, $\alpha$ and $n$ are the frequency-dependent absorption and refractive index, $L$ is the cavity length, and $\phi$ is the phase accumulated upon reflection against a mirror. $T$ and $R$ are the cavity transmittance and reflectance, respectively, where it is assumed that $T = 1 - R$.

Models.dielectric_real — Method

Frequency-dependent real part of the dielectric function in terms of the background index and Lorentzian oscillator,

$\varepsilon_1(\nu) = n^2 + \frac{A (\nu_0^2 - \nu^2)}{(\nu^2 - \nu_0^2)^2 + (\Gamma\nu)^2},$

where $n$ is the background real index, $\nu_0$ is the excitation frequency, $\Gamma$ is the line width of the oscillator, and $A$ is the oscillator strength.

Models.dielectric_imag — Method

Frequency-dependent imaginary part of the dielectric function.

$\varepsilon_2(\nu) = \frac{A \Gamma \nu}{(\nu^2 - \nu_0^2)^2 + (\Gamma\nu)^2}$

Models.cavity_mode_energy — Method

Cavity mode energy as a function of incident angle.

$E_\text{cavity}(\theta) = E_0 \left( 1 - \frac{\sin^2(\theta)}{n^2} \right)^{-1/2},$

where $E_0$ is the energy of the cavity mode at zero degrees incidence angle.

Models.coupled_energies — Function

Coupled energies found by diagonalizing the coupled-oscillator model Hamiltonian:

$H = \begin{pmatrix} E_c(\theta) & V/2 \\V/2 & E_v \end{pmatrix}.$

Diagonalizing $H$ gives the interaction energies as a function of $\theta$. These exhibit avoided crossing behavior near where the cavity mode energy equals the material excitation energy.

$V_{\pm}(\theta) = \frac{1}{2}(E_c(\theta) + E_v) \pm \frac{1}{2} \sqrt{(E_c(\theta) - E_v)^2 + V^2}$

https://en.wikipedia.org/wiki/Avoided_crossing

Models.fsr — Method

Find the free spectral range for all adjacent peaks. Return the list of FSRs, the average FSR for the range, and the standard deviation.

Models.fsr — Method

Solve for one of three variables in the equation for free spectral range in terms of the other two.

$\Delta \nu = \frac{1}{2 L n},$

where $\Delta\nu = |\nu_2 - \nu_2|$, $L$ is the intracavity length, and $n$ is intracavity index of refraction.