Special Functions

These function are specific to cavity polariton research, where things like the cavity transmittance, material dielectric function, and cavity free spectral range (FSR) need to be calculated and modeled.

Models.cavity_mode_energy — Method

cavity_mode_energy(θ::Real, p)

p = [E_0, n_eff]

Cavity mode energy as a function of incident angle.

\[\begin{aligned} E_\text{cavity}(\theta) = E_0 \left( 1 - \frac{\sin^2(\theta)}{n^2} \right)^{-1/2}, \end{aligned}\]

where $E_0$ is the energy of the cavity mode at zero degrees incidence angle.

Models.cavity_transmittance — Method

cavity_transmittance(ν, p)

p = [n, α, L, T, R, ϕ]

Cavity transmittance as a function of frequency.

\[\begin{aligned} T(\nu) = \frac{T^2 e^{-\alpha L}}{1 + R^2 e^{-2 \alpha L} - 2 R e^{-\alpha L} \cos(4\pi n L \nu + 2\phi)}, \end{aligned}\]

where $\nu$ is the frequency, $\alpha$ and $n$ are the frequency-dependent absorption and refractive index, $L$ is the cavity length, and $\phi$ is the phase accumulated upon reflection against a mirror. $T$ and $R$ are the cavity transmittance and reflectance, respectively, where it is assumed that $T = 1 - R$.

Models.coupled_energies — Method

coupled_energies(E_c::Vector{Real}, E_v::Real, V::Real, branch=0)

Coupled energies found by diagonalizing the coupled-oscillator model Hamiltonian with principle energies E_c (cavity photon) and E_v (material excitation), and interaction energy V. The lower and upper branches are called with branch=0 and branch=1, respectively.

\[\begin{aligned} H = \begin{pmatrix} E_1(\theta) & \Omega_R/2 \\\Omega_R/2 & E_2 \end{pmatrix} \end{aligned}\]

Diagonalizing $H$ gives the interaction energies as a function of $\theta$. These exhibit avoided crossing behavior near where the cavity mode energy equals the material excitation energy.

\[\begin{aligned} E_{\pm}(\theta) = \frac{1}{2}(E_1(\theta) + E_2) \pm \frac{1}{2} \sqrt{(E_1(\theta) - E_2)^2 + \Omega_R^2} \end{aligned}\]

https://en.wikipedia.org/wiki/Avoided_crossing

Models.dielectric_imag — Method

dielectric_imag(ν, p)

p = [ν_0, A, Γ, y_0]

Frequency-dependent imaginary part of the dielectric function.

\[\begin{aligned} \varepsilon_2(\nu) = \frac{A \Gamma \nu}{(\nu^2 - \nu_0^2)^2 + (\Gamma\nu)^2} \end{aligned}\]

Models.dielectric_real — Method

dielectric_real(ν, p)

p = [ν_0, A, Γ, n_eff]

Frequency-dependent real part of the dielectric function in terms of the background index and Lorentzian oscillator,

\[\begin{aligned} \varepsilon_1(\nu) = n^2 + \frac{A (\nu_0^2 - \nu^2)}{(\nu^2 - \nu_0^2)^2 + (\Gamma\nu)^2}, \end{aligned}\]

where $n$ is the background real index, $\nu_0$ is the excitation frequency, $\Gamma$ is the line width of the oscillator, and $A$ is the oscillator strength.

Models.fsr — Method

fsr(peak_positions::Vector{Real)

Find the free spectral range for all adjacent peaks. Return the list of FSRs, the average FSR for the range, and the standard deviation.

Models.fsr — Method

fsr(x1::Real, x2::Real)

Solve for one of three variables in the equation for free spectral range in terms of the other two.

\[\begin{aligned} \Delta \nu = \frac{1}{2 L n}, \end{aligned}\]

where $\Delta\nu = |\nu_2 - \nu_1|$, $L$ is the intracavity length, and $n$ is intracavity index of refraction.